+++ NEWS.md
@@ -27,6 +27,9 @@ New library functions
 New library features
 --------------------
 
+* `tempname` can now take a suffix string to allow the file name to include a suffix and include that suffix in
+  the uniquing checking ([#])
+
 Standard library changes
 ------------------------